Test for TextValidator
Now there's a test running on android junit that tests validation of EditTexts.
Also, I changed the program so that the build is not environment specific (it doesn't need to run only on machines that I control anymore).

In order to build intern now, though, the user will have to create a settings.xml file with the template:

<settings xmlns="http://maven.apache.org/SETTINGS/1.0.0"
  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
  xsi:schemaLocation="http://maven.apache.org/SETTINGS/1.0.0
                      http://maven.apache.org/xsd/settings-1.0.0.xsd">
  <profiles>
    <profile>
      <id>dev</id>
      <activation>
	<activeByDefault>true</activeByDefault>
      </activation>
      <properties>
	<android.sdk.home>type your sdk location here</android.sdk.home>
      </properties>
    </profile>
  </profiles>

and the project will (hopefully) build after that.