Testfix: Not finding test package on jenkins.
Turns out this apklib doesn't solve some problems. The test application need an application to test, it cannot be a lib. So I created a stub application that depends on the apklib (and on other apklibs, should they appear on the future) in such a way that I can test them.

This also will make it simpler to identify errors such as the one that happened on Kotoba (transitive dependency errors), since the structure of intern and kotoba are now the same.